Separated the Counts array into an array with just the count values and another array to hold the Escape Velocities, both of type uShort.

+++ src/MSS.Types/Operation/HistogramALow.cs
@@ -31,7 +31,7 @@ namespace MSS.Types
 		//	}
 		//}
 
-		public HistogramALow(IEnumerable<int> values)
+		public HistogramALow(IEnumerable<ushort> values)
 		{
 			var low = values.Min();
 			var high = values.Max();
